Single_data_input

Input the Single data in seruat format

Single_data_input(
  Pagwas,
  Single_data,
  nfeatures = NULL,
  FilterSingleCell = FALSE,
  Pathway_list = NULL,
  min.lib.size = 1000,
  min.reads = 10,
  min.detected = 5,
  min_clustercells = 50
)

Arguments

Pagwas

Pagwas format

Single_data

Input the Single data in seruat format, Idents should be the celltypes annotation.

nfeatures

The parameter for FindVariableFeatures, NULL means select all genes

FilterSingleCell

whther to filter the single cell data,if you filter it before,choose FALSE, otherwise set TRUE.

Pathway_list

(list,character) pathway gene sets list

min.lib.size

Threshold for single data library.

min.reads

Threshold for total reads fo each cells.

min.detected

Threshold for total cells fo each gene.

min_clustercells

Threshold for total cells fo each cluster.

Value

Pagwas

Examples

library(scPagwas)
data(scRNAexample)
# Pagwas is better getting from GWAS_summary_input(),or NULL is right,
Pagwas <- Single_data_input(Pagwas = Pagwas, Single_data = scRNAexample)
#> Input single cell count data and cell annotation data!
#> Error in Single_data_input(Pagwas = Pagwas, Single_data = scRNAexample): object 'Pagwas' not found