merge_pagwas

Merging several pagwas result to one result.When your single cell data is too large to run, you can split your data into several split, then use this function to merge them. Split the data randomly or in order of cellytpes.

Usage

merge_pagwas(
  Single_data = NULL,
  n_topgenes = 1000,
  output.dirs = "splice_scPagwas",
  assay = "RNA",
  random = FALSE,
  seed = 1234,
  random_times = 10,
  proportion = 0.3
)

Arguments

Single_data

The Whole single cell data set. rds files in seurat format. The filesnames or seurat format data.

n_topgenes

(integr)Number of top associated gene selected to calculate the scPagwas score;

output.dirs

The files names for the result of scPagwas_main function.

assay

The assay used for single cell data.

random

Whether split the Single_data in random. Using id when the Single_data is too big to run.

seed

The seed for random.

random_times

The times for random, default is 10.

proportion

The proportion for spliting the Single_data; the bigger for Single_data the less proportion for spliting, and the more times for random_times.

Value

seurat format result for scPagwas.

Note

1.The rownames genes and gwas data must uniform. 2.The merge_pagwas function only can use to merge the seurat format result. That is "singlecell = T;celltype =F" can use to all the split data, "singlecell = T;celltype = T" can be used only when the data is splited by celltypes. but "singlecell = F;celltype =T" can not be used.

Author

Chunyu Deng

Examples

library(Seurat)
#> Attaching SeuratObject
scRNAexample <-readRDS(system.file("extdata", "scRNAexample.rds", package = "scPagwas"))
#set the output.dirs for any directory.
set.dirs="D:/OneDrive/GWAS_Multiomics/Pagwas/test/"
#set the number of split time, it depend on the size of your single cell data.
#There have two form of spliting the data:

#########
##1.By Random
#Create the random index number.
n_split=2
Split_index <- rep(1:n_split, time = ceiling(ncol(scRNAexample)/n_split), length = ncol(scRNAexample))

for (i in 1:n_split) {
  Example_splice <- scRNAexample[,Split_index==i]
  saveRDS(Example_splice,file = paste0(set.dirs,"Example_splice",i,".rds"))
}

#########
###1.By celltypes
#Create the random index number.
for (i in levels(Idents(scRNAexample))) {
  Example_splice <- subset(scRNAexample,idents = i)
  saveRDS(Example_splice,file = paste0(set.dirs,"Example_splice",i,".rds"))
}
Pagwas<-list()
gwas_data <- bigreadr::fread2(system.file("extdata", "GWAS_summ_example.txt", package = "scPagwas"))
Pagwas <- GWAS_summary_input(
  Pagwas = Pagwas,
  gwas_data = gwas_data,
  maf_filter = 0.1
)
#> Input gwas summary data frame!
#> No "chr" from chrom!, now pasting it!
#> Filtering out SNPs with MAF criterion
Pagwas$snp_gene_df <- Snp2Gene(snp = Pagwas$gwas_data, refGene = block_annotation, marg = 10000)
#> Error in Snp2Gene(snp = Pagwas$gwas_data, refGene = block_annotation,     marg = 10000): could not find function "Snp2Gene"
names(Pagwas)
#> [1] "gwas_data"
setwd(set.dirs)

for (i in 1:n_split) {
  scPagwas_main(Pagwas =Pagwas,
                gwas_data =NULL,
                Single_data = paste0("Example_splice",i,".rds"), #the
                output.prefix=i, #the prefix can be the circulation coefficient i.
                output.dirs="splice_scPagwas", #the output.dirs shoukd be the same for each circulation
                Pathway_list=Genes_by_pathway_kegg,
                run_split=TRUE, #This parameter is set to run single result for each split result.
                min_clustercells=1, #the minimum cluster cell number should be 1.
                assay="RNA",
                block_annotation = block_annotation,
                chrom_ld = chrom_ld)
}
#> ##------ Fri Mar 10 15:50:21 2023 ------## ******* 1st: Single_data_input function start! ********
#> ** Start to read the single cell data!
#> ** Start to filter single cell data!
#> Input single cell data!
#> 10 cell types are remain, after filter!
#> done!
#> ##------ Fri Mar 10 15:50:22 2023 ------## ******* 2nd: Pathway_pcascore_run function start!! ********
#> * Start to get Pathway SVD socre!
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%[1] "Celltype1"
#>   |                                                                              |==============                                                        |  20%[1] "Celltype3"
#>   |                                                                              |=====================                                                 |  30%[1] "Celltype5"
#>   |                                                                              |============================                                          |  40%[1] "Celltype4"
#>   |                                                                              |===================================                                   |  50%[1] "Celltype9"
#>   |                                                                              |==========================================                            |  60%[1] "Celltype7"
#>   |                                                                              |=================================================                     |  70%[1] "Celltype10"
#>   |                                                                              |========================================================              |  80%[1] "Celltype8"
#>   |                                                                              |===============================================================       |  90%[1] "Celltype2"
#>   |                                                                              |======================================================================| 100%[1] "Celltype6"
#> 
#> Using score as value column: use value.var to override.
#> done!
#> ##------ Fri Mar 10 15:50:38 2023 ------## ******* 3rd: GWAS_summary_input function start! ********
#> Error in scPagwas_main(Pagwas = Pagwas, gwas_data = NULL, Single_data = paste0("Example_splice",     i, ".rds"), output.prefix = i, output.dirs = "splice_scPagwas",     Pathway_list = Genes_by_pathway_kegg, run_split = TRUE, min_clustercells = 1,     assay = "RNA", block_annotation = block_annotation, chrom_ld = chrom_ld): Error: gwas_data should be input!
if(length(SOAR::Objects())>0){
  SOAR::Remove(SOAR::Objects())
}

setwd(set.dirs) #you must set the workfiles the same as before
Pagwas_integrate <- merge_pagwas(Single_data = system.file("extdata", "scRNAexample.rds", package = "scPagwas") ,# read the whole single cell data.
                                output.dirs="splice_scPagwas", #The same with scPagwas_main functionn_topgenes = 1000,
                                assay='RNA',
                                random=T,
                                seed=1234,
                                random_times=10,
                                proportion=0.3)
#> ** Start to read the single cell data!
#> Warning: The total length of the result of scPagwas.gPAS.score for your splited single cell data
#>             is not the same with your Single_data, Please check the code before or the min_clustercells
#>             parameter(set to 1)
#> Start to run the corSparse function in random!
#> [1] "Randome Times: 1"
#> [1] "Randome Times: 2"
#> [1] "Randome Times: 3"
#> [1] "Randome Times: 4"
#> [1] "Randome Times: 5"
#> [1] "Randome Times: 6"
#> [1] "Randome Times: 7"
#> [1] "Randome Times: 8"
#> [1] "Randome Times: 9"
#> [1] "Randome Times: 10"
#> Start to run the scGene_rankP function in random!
#> [1] "Randome Times: 1"
#> Error in `[.data.frame`(CellScalepValue, , "pValueHighScale"): undefined columns selected