plot_vln_Corgenes

Make grid of violin plots Thanks to the SCOPfunctions package https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R

produce a n_celltype * n_genes grid of violin plots

plot_vln_Corgenes(
  seurat_obj,
  assay,
  slot,
  var_group,
  vec_features,
  vec_group_colors = NULL,
  f_color = colorRampPalette(brewer.pal(n = 11, name = "RdYlBu")),
  flip = T,
  do_plot = F,
  pt.size = 0,
  feature_fontface = "bold.italic",
  fontsize_axistext_x = 12,
  fontsize_axistext_y = 12,
  aspect.ratio = NULL,
  figurenames = NULL,
  width = 7,
  height = 7
)

Arguments

seurat_obj

Seurat object (Seurat ^3.0)

assay

seurat_obj assay to use

slot

seurat_obj slot to use

var_group

the group variable, character

vec_features

a vector of features to plot in the violin plot

vec_group_colors

a vector of colors, named by corresponding group. Length must match number of groups. Character

f_color

if vec_group_colors is not provided, the user may instead provide a function f_color() that takes as its only argument the number of colors to generate

flip

if TRUE (default), groups are rows and features are columns, and vice-verso for FALSE

do_plot

Whether to plot, logical

pt.size

size of jitter in the violin plots. Set to 0 (default) to omit

figurenames

The filename and address of the output plot,

width

figure width

height

figure height

Value

ggplot2 object

Examples

top5genes <- rownames(Pagwas$gene_heritability_correlation)[order(Pagwas$gene_heritability_correlation, decreasing = T)[1:5]]
#> Error in rownames(Pagwas$gene_heritability_correlation): object 'Pagwas' not found
library("RColorBrewer")
library("ggplot2")

plot_vln_Corgenes(
  seurat_obj = scRNAexample,
  assay = "RNA", slot = "data",
  var_group = "anno", # 缂佸棜鍎猚luster濞夈劑鍣撮崚悚笺阜NA= top5genes,
  vec_group_colors = pal_d3(alpha = 0.5)(10)
)
#> Error in plot_vln_Corgenes(seurat_obj = scRNAexample, assay = "RNA", slot = "data",     var_group = "anno", vec_features = top5genes, vec_group_colors = pal_d3(alpha = 0.5)(10)): object 'scRNAexample' not found