Main wrapper functions for scPagwas

Main Pagwas wrapper re-progress function.

scPagwas_main(
  Pagwas = NULL,
  gwas_data = NULL,
  output.prefix = "Test",
  output.dirs = "scPagwastest_output",
  block_annotation = NULL,
  Single_data = NULL,
  assay = c("RNA", "SCT"),
  Pathway_list = NULL,
  chrom_ld = NULL,
  marg = 10000,
  maf_filter = 0.01,
  min_clustercells = 10,
  min.pathway.size = 5,
  max.pathway.size = 300,
  iters = 200,
  n_topgenes = 1000,
  singlecell = T,
  celltype = T,
  seruat_return = T,
  remove_outlier = T,
  ncores = 1
)

Arguments

Pagwas

(list)default is "NULL" when you first run the funciton; Pagwas should be list class; Sometimes, It can inherit the result from "scPagwas_main" function last time, when you turn the "seruat_return" to FALSE; It is suitable for circulation running for the same single data.

gwas_data

(data.frame)GWAS summary data; It must have some colmuns such as: chrom| pos | rsid | se | beta | maf 6 | 119968580 | rs1159767 | 0.032 | 0.019 |0.5275 10 | 130566523 | rs559109 | 0.033 | 0.045 |0.4047 5 | 133328825 | rs6893145 | 0.048 | 0.144 |0.1222 7 | 146652932 | rs13228798| 0.035 | 0.003 | 0.3211

output.prefix

(character)output.prefix,prefix used for output tables;

output.dirs

(character)output directory for result; If the directory is nonexistence, the function will create it;

block_annotation

(data.frame) Start and end points for block traits, usually genes.

Single_data

(character or Seruat)Input the Single data in seruat format, or the Seruat data address for rds format.Idents should be the celltypes annotation.

assay

(character)assay data of your single cell data to use,default is "RNA"

Pathway_list

(list,character) pathway gene sets list

chrom_ld

(list,numeric)LD data for 22 chromosome.

marg

(integr) the distance to TSS site,default is 10000, then gene-TSS-window size is 20000.

maf_filter

(numeric)Filter the maf, default is 0.01

min_clustercells

(integr)Only use is when FilterSingleCell is TRUE.Threshold for total cells fo each cluster.default is 10

min.pathway.size

(integr)Threshold for min pathway gene size. default is 5

max.pathway.size

(integr)Threshold for max pathway gene size. default is 300

iters

(integr)number of bootstrap iterations to perform

n_topgenes

(integr)Number of top associated gene selected to calculate the scPagwas score;

singlecell

(logical)Whether to produce the singlecell result;

celltype

(logical)Whether to produce the celltypes result;

seruat_return

(logical) Whether return the Seruat format result, if not,will return a list result;

remove_outlier

(logical)Whether to remove the outlier for scPagwas score.

ncores

(integr)Parallel cores,default is 1. use detectCores() to check the cores in computer.

Value

Returns a Seruat data with entries(seruat_return=T):

assay:

scPagwasPaPca:An assay for S4 type of data; the svd result for pathways in each cells; scPagwasPaHeritability:An assay for S4 type of data; the gPas matrix for pathways in each cells;

meta.data

scPagwas.topgenes.Score1: the column for "meta.data";Enrichment socre for inheritance associated top genes. scPagwas.gPAS.score: the column for "meta.data";Inheritance regression effects for each cells CellScalepValue: the column for "meta.data";CellScalepValue for each cells CellScaleqValue: the column for "meta.data";CellScaleqValue for each cells

misc: element in result,Pagwas@misc

Pathway_list:a list for pathway gene list intersecting with single cell data pca_cell_df: a data frame for pathway pca result for each celltype. lm_results: the regression result for each cell. gene_heritability_correlation: heritability correlation value for each gene; scPathways_rankPvalue:pvaues for each pathway bootstrap_results:The bootstrap data frame results for celltypes including bootstrap pvalue and confidence interval.

Returns files:

scPagwas.run.log: the running log file for scPagwas _parameters.txt:parameters log file for scPagwas _pca_scCell_mat.txt: 1st svd result for pathways in each cell _pca_celltypes_mat.txt: 1st svd results for pathways in celltypes _singlecell_scPagwas_score.Result.csv: The final result for lm and top gene score _celltypes_bootstrap_results.csv:The bootstrap data frame results for celltypes including bootstrap pvalue and confidence interval _gene_heritability_correlation.csv: heritability correlation value("cor" for pearson) for each gene;

Returns a list class with entries(seruat_return=F):

scPagwasPaPca:Assays for S4 type of data; the svd result for pathways in each cells; scPagwas.topgenes.Score1: the column for "meta.data";Enrichment socre for inheritance associated top genes. sclm_score: the column for "meta.data";Inheritance regression effects for each cells Pathway_list: The number of Lanczos iterations carried out pca_cell_df: The total number of matrix vector products carried out sclm_results: The total number of matrix vector products carried out allsnp_gene_heritability_correlation: The total number of matrix vector products carried out Pathway_ctlm_results: The total number of matrix vector products carried out lm_results: The total number of matrix vector products carried out Pathway_ct_results: The total number of matrix vector products carried out

Details

scPagwas, A single-cell pathway-based principal component (PC)-scoring algorithm named for integrating polygenic signals from GWAS with single cell data to infer disease-relevant cell populations and score the associations of individual cells with complex diseases.

The entry point for Pagwas analysis.

Note

1.When you run the package in linux server, you can run export OPENBLAS_NUM_THREADS=1 before enter into R environment, It can help you to avoid core errors. 2. When the code is kill for some wrong, you can run SOAR::Objects() to check the objects, or run SOAR::Remove(SOAR::Objects()) to remove it. 3. When your data is too big to run the function(bus error), you can split your data into several sub-data and run the merge_pagwas to merge them.

Examples

library(scPagwas)
library(ggplot2)
library(Seurat)
#> Attaching SeuratObject
Pagwas_data <- scPagwas_main(
  Pagwas = NULL,
  gwas_data = system.file("extdata", "GWAS_summ_example.txt", package = "scPagwas"),
  Single_data = system.file("extdata", "scRNAexample.rds", package = "scPagwas"),
  output.prefix = "test",
  output.dirs = "scPagwastest_output",
  block_annotation = block_annotation,
  assay = "RNA",
  Pathway_list = Genes_by_pathway_kegg,
  chrom_ld = chrom_ld,
  singlecell = T,
  celltype = T,
  ncores = 1
)
#> ##------ Wed Jul 06 17:48:14 2022 ------## ******* 1st: Single_data_input function start! ********
#> ** Start to read the single cell data!
#> Input single cell data!
#> 10 cell types are remain, after filter!
#> done!
#> ##------ Wed Jul 06 17:48:18 2022 ------## ******* 2nd: Pathway_pcascore_run function start!! ********
#> * Start to get Pathway SVD socre!
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%[1] "Celltype1"
#>   |                                                                              |==============                                                        |  20%[1] "Celltype2"
#>   |                                                                              |=====================                                                 |  30%[1] "Celltype3"
#>   |                                                                              |============================                                          |  40%[1] "Celltype4"
#>   |                                                                              |===================================                                   |  50%[1] "Celltype5"
#>   |                                                                              |==========================================                            |  60%[1] "Celltype6"
#>   |                                                                              |=================================================                     |  70%[1] "Celltype7"
#>   |                                                                              |========================================================              |  80%[1] "Celltype8"
#>   |                                                                              |===============================================================       |  90%[1] "Celltype9"
#>   |                                                                              |======================================================================| 100%[1] "Celltype10"
#> 
#> Using score as value column: use value.var to override.
#> done!
#> ##------ Wed Jul 06 17:48:25 2022 ------## ******* 3rd: GWAS_summary_input function start! ********
#> ** Start to read the gwas_data!
#> Input gwas summary data frame!
#> No "chr" from chrom!, now pasting it!
#> Filtering out SNPs with MAF criterion
#> done!
#> ##------ Wed Jul 06 17:48:25 2022 ------## ******* 4th: Snp2Gene start!! ********
#> * Computing distances between SNPs and genes
#> 6604 SNPs not inside gene domain
#> ** Computing distance to domain starts
#> ** Computing distance to domain ends
#> ##------ Wed Jul 06 17:48:25 2022 ------## ******* 5th: Pathway_annotation_input function start! ********
#> obtain the pathway block information
#> done!
#> ##------ Wed Jul 06 17:48:26 2022 ------## ******* 6th: Link_pathway_blocks_gwas function start! ********
#> * Start to link gwas and pathway block annotations for 281 pathways!
#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |===                                                                   |   5%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |=======                                                               |  11%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |==============                                                        |  21%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |=================                                                     |  25%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |=====================                                                 |  31%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |============================                                          |  41%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |===============================                                       |  45%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%
#> done!
#> ##------ Wed Jul 06 17:48:54 2022 ------## ******* 7th: Celltype_heritability_contributions function start! ********
#> ** Start inference
#> * starting bootstrap iteration for 200 times
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |==                                                                    |   4%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |=====                                                                 |   8%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |=========                                                             |  14%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  16%  |                                                                              |============                                                          |  17%  |                                                                              |============                                                          |  18%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |================                                                      |  24%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |===================                                                   |  28%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |=======================                                               |  34%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |==========================                                            |  38%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |==============================                                        |  44%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |=================================                                     |  48%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  62%  |                                                                              |============================================                          |  63%  |                                                                              |============================================                          |  64%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |===================================================                   |  74%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |==========================================================            |  84%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |=================================================================     |  94%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |====================================================================  |  98%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> done!
#> ##------ Wed Jul 06 17:48:55 2022 ------## ******* 8th: scPagwas_perform_score function start! ********
#> * Get Pathways'rankPvalue for each celltypes!
#> * Get scPgwas score for each single cell
#> done!
#> ##------ Wed Jul 06 17:49:01 2022 ------## ******* 9th: scGet_gene_heritability_correlation function start! ********
#> done
#> * Get rankPvalue for each single cell