Estimate local False Discovery Rate (FDR)

Estimate the local FDR values from p-values.

Usage

lfdr(
  p,
  pi0 = NULL,
  trunc = TRUE,
  monotone = TRUE,
  transf = c("probit", "logit"),
  adj = 1.5,
  eps = 10^-8,
  ...
)

Arguments

p

A vector of p-values (only necessary input).

pi0

Estimated proportion of true null p-values.

trunc

If TRUE, local FDR values >1 are set to 1. Default is TRUE.

monotone

If TRUE, local FDR values are non-decreasing with increasing p-values. Default is TRUE; this is recommended.

transf

Either a "probit" or "logit" transformation is applied to the p-values so that a local FDR estimate can be formed that does not involve edge effects of the c(0,1) interval in which the p-values lie.

adj

Numeric value that is applied as a multiple of the smoothing bandwidth used in the density estimation. Default is adj=1.0.

eps

Numeric value that is threshold for the tails of the empirical p-value distribution. Default is 10^-8.

...

Additional arguments.

Value

A vector of estimated local FDR values, with each entry corresponding to the entries of the input p-value vector p.

References

Efron B, Tibshirani R, Storey JD, and Tisher V. (2001) Empirical Bayes analysis of a microarray experiment. Journal of the American Statistical Association, 96: 1151-1160.
http://www.tandfonline.com/doi/abs/10.1198/016214501753382129

Storey JD. (2003) The positive false discovery rate: A Bayesian interpretation and the q-value. Annals of Statistics, 31: 2013-2035.
http://projecteuclid.org/DPubS/Repository/1.0/Disseminate?view=body&id=pdf_1&handle=euclid.aos/1074290335

Storey JD. (2011) False discovery rates. In International Encyclopedia of Statistical Science.
http://genomine.org/papers/Storey_FDR_2011.pdf
http://www.springer.com/statistics/book/978-3-642-04897-5

Author

John D. Storey