merge_pagwas

Merging several pagwas result to one result.When your single cell data is too large to run, you can split your data into several split, then use this function to merge them. Split the data randomly or in order of cellytpes.

merge_pagwas(Pagwas_list = NULL, n_topgenes = 1000)

Arguments

Pagwas_list

list format, element including seruat format result.

n_topgenes

(integr)Number of top associated gene selected to calculate the scPagwas score;

Value

Note

1.The rownames genes and gwas data must uniform. 2.The merge_pagwas function only can use to merge the seruat format result. That is "singlecell = T;celltype =F" can use to all the split data, "singlecell = T;celltype = T" can be used only when the data is splited by celltypes. but "singlecell = F;celltype =T" can not be used.

Author

Chunyu Deng

Examples

library(scPagwas)
Pagwas<-list()
library(Seurat)
#> Attaching SeuratObject
scRNAexample <-readRDS(system.file("extdata", "scRNAexample.rds", package = "scPagwas"))
Example_splice1<-subset(scRNAexample,idents=c("Celltype1","Celltype2","Celltype3","Celltype4", "Celltype5"))
Example_splice2<-subset(scRNAexample,idents=c( "Celltype6","Celltype7","Celltype8","Celltype9","Celltype10" ))

gwas_data <- bigreadr::fread2(system.file("extdata", "GWAS_summ_example.txt", package = "scPagwas"))
Pagwas <- GWAS_summary_input(
  Pagwas = Pagwas,
  gwas_data = gwas_data,
  maf_filter = 0.1
)
#> Input gwas summary data frame!
#> No "chr" from chrom!, now pasting it!
#> Filtering out SNPs with MAF criterion
Pagwas$snp_gene_df <- Snp2Gene(snp = Pagwas$gwas_data, refGene = block_annotation, marg = 10000)
#> * Computing distances between SNPs and genes
#> 5332 SNPs not inside gene domain
#> ** Computing distance to domain starts
#> ** Computing distance to domain ends
#Run the first split data
Pagwas1<-scPagwas_main(Pagwas =Pagwas,
                      gwas_data =NULL,
                       Single_data =Example_splice1,
                       output.prefix="splice1",
                       output.dirs="splice_scPagwas",
                       Pathway_list=Genes_by_pathway_kegg,
                       ncores=5,
                       assay="RNA",
                       block_annotation = block_annotation,
                       chrom_ld = chrom_ld)
#> ##------ Tue Jul 19 14:15:20 2022 ------## ******* 1st: Single_data_input function start! ********
#> Input single cell data!
#> 5 cell types are remain, after filter!
#> done!
#> ##------ Tue Jul 19 14:15:24 2022 ------## ******* 2nd: Pathway_pcascore_run function start!! ********
#> * Start to get Pathway SVD socre!
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%[1] "Celltype1"
#>   |                                                                              |============================                                          |  40%[1] "Celltype2"
#>   |                                                                              |==========================================                            |  60%[1] "Celltype3"
#>   |                                                                              |========================================================              |  80%[1] "Celltype4"
#>   |                                                                              |======================================================================| 100%[1] "Celltype5"
#> 
#> Using score as value column: use value.var to override.
#> done!
#> ##------ Tue Jul 19 14:15:31 2022 ------## ******* 3rd: GWAS_summary_input function start! ********
#> ##------ Tue Jul 19 14:15:31 2022 ------## ******* 5th: Pathway_annotation_input function start! ********
#> obtain the pathway block information
#> done!
#> ##------ Tue Jul 19 14:15:33 2022 ------## ******* 6th: Link_pathway_blocks_gwas function start! ********
#> * Start to link gwas and pathway block annotations for 278 pathways!
#>   |                                                                              |                                                                      |   0%
#> Error in setkeyv(x, cols, verbose = verbose, physical = physical): x is not a data.table
Pagwas2<-scPagwas_main(Pagwas =Pagwas,
                      gwas_data =NULL,
                      Single_data =Example_splice2,
                      output.prefix="splice2 ",
                      output.dirs="splice_scPagwas",
                      Pathway_list=Genes_by_pathway_kegg,
                      ncores=5,
                      assay="RNA",
                      block_annotation = block_annotation,
                      chrom_ld = chrom_ld)
#> ##------ Tue Jul 19 14:15:33 2022 ------## ******* 1st: Single_data_input function start! ********
#> Input single cell data!
#> 5 cell types are remain, after filter!
#> done!
#> ##------ Tue Jul 19 14:15:34 2022 ------## ******* 2nd: Pathway_pcascore_run function start!! ********
#> * Start to get Pathway SVD socre!
#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%[1] "Celltype6"
#>   |                                                                              |============================                                          |  40%[1] "Celltype7"
#>   |                                                                              |==========================================                            |  60%[1] "Celltype8"
#>   |                                                                              |========================================================              |  80%[1] "Celltype9"
#>   |                                                                              |======================================================================| 100%[1] "Celltype10"
#> 
#> Using score as value column: use value.var to override.
#> done!
#> ##------ Tue Jul 19 14:15:39 2022 ------## ******* 3rd: GWAS_summary_input function start! ********
#> ##------ Tue Jul 19 14:15:39 2022 ------## ******* 5th: Pathway_annotation_input function start! ********
#> obtain the pathway block information
#> done!
#> ##------ Tue Jul 19 14:15:40 2022 ------## ******* 6th: Link_pathway_blocks_gwas function start! ********
#> * Start to link gwas and pathway block annotations for 273 pathways!
#>   |                                                                              |                                                                      |   0%
#> Error in setkeyv(x, cols, verbose = verbose, physical = physical): x is not a data.table
Pagwas_integrate <- merge_pagwas(Pagwas_list = list(Pagwas1,Pagwas2),n_topgenes = 1000)
#> Error in merge_pagwas(Pagwas_list = list(Pagwas1, Pagwas2), n_topgenes = 1000): object 'Pagwas1' not found