plot_bar_positie_nagtive

Generate barplot of identity group composition Thanks to the SCOPfunctions package https://github.com/CBMR-Single-Cell-Omics-Platform/SCOPfunctions/blob/main/R/plot.R Generate a percentage barplot that shows the composition of each identity (e.g. sample)in terms of groups (e.g. positive and negative cells for scPagwas)

plot_bar_positie_nagtive(
  seurat_obj,
  var_ident,
  var_group,
  vec_group_colors = NULL,
  f_color = colorRampPalette(brewer.pal(n = 11, name = "RdYlBu")),
  do_plot = F,
  title = NULL,
  p_thre = 0.05,
  fontsize_title = 24,
  fontsize_axistitle_x = 18,
  fontsize_axistitle_y = 18,
  fontsize_axistext_x = 12,
  fontsize_axistext_y = 12,
  fontsize_legendtitle = 12,
  fontsize_legendtext = 10,
  aspect.ratio = 1.2,
  output.prefix = "Test",
  output.dirs = NULL,
  width = 7,
  height = 7
)

Arguments

seurat_obj

Seurat object (Seurat ^3.0)

var_ident

the identify variable, character

var_group

the group variable, character

vec_group_colors

a vector of colors, named by corresponding group. Length must match number of groups. Character

f_color

if vec_group_colors is not provided, the user may instead provide a function f_color() that takes as its only argument the number of colors to generate

do_plot

Whether to plot, logical

title

NULL to leave out

fontsize_title

NULL to leave out

fontsize_axistitle_x

NULL to leave out

fontsize_axistitle_y

NULL to leave out

fontsize_axistext_x

NULL to leave out

fontsize_axistext_y

NULL to leave out

fontsize_legendtitle

NULL to leave out

fontsize_legendtext

NULL to leave out

width

figure width

height

figure height

figurenames

The filename and address of the output plot,

Value

a ggplot2 object

Examples

p <- plot_barIdentGroup(seurat_obj = seu, var_ident = "sample", var_group = "cluster")
#> Error in plot_barIdentGroup(seurat_obj = seu, var_ident = "sample", var_group = "cluster"): could not find function "plot_barIdentGroup"